// scripts/interfaces/Spec.ts
export enum Spec {
    havoc = "havoc",
    vengeance = "vengeance",
    blood = "blood",
    unholy = "unholy",
    balance = "balance",
    feral = "feral",
    guardian = "guardian",
    beast_mastery = "beast-mastery",
    marksmanship = "marksmanship",
    survival = "survival",
    arcane = "arcane",
    fire = "fire",
    frost = "frost",
    brewmaster = "brewmaster",
    windwalker = "windwalker",
    mistweaver = "mistweaver",
    holy = "holy",
    retribution = "retribution",
    discipline = "discipline",
    shadow = "shadow",
    assassination = "assassination",
    outlaw = "outlaw",
    subtlety = "subtlety",
    elemental = "elemental",
    enhancement = "enhancement",
    restoration = "restoration",
    affliction = "affliction",
    demonology = "demonology",
    destruction = "destruction",
    arms = "arms",
    fury = "fury",
    protection = "protection",
}